refactoring du code
