+++ src/main.py
@@ -1,95 +1,56 @@
 import sys
+import os
+from decrypt_mess1 import brute_force_caesar
+from decrypt_mess2 import vigenere_decrypt, read_ciphertext
+from decrypt_mess3 import fon, adfgvx
 
-def caesar_decrypt(ciphertext, shift):
+def detect_cipher_type(file_name):
     """
-    Déchiffre un message crypté avec le chiffrement de César en fonction du décalage fourni.
-    Seules les lettres majuscules sont déchiffrées, les autres caractères restent inchangés.
+    Détecte le type de chiffrement à partir du nom de fichier.
+    Si le fichier contient 'message1', on suppose que c'est du César,
+    'message2' pour Vigenère, et 'message3' pour ADFGVX.
     """
-    decrypted_text = ''
-    for char in ciphertext:
-        if char.isalpha() and char.isupper():
-            shift_base = ord('A') 
-            decrypted_char = chr((ord(char) - shift_base - shift) % 26 + shift_base)
-            decrypted_text += decrypted_char
-        else:
-            decrypted_text += char
-    return decrypted_text
-
-def brute_force_caesar(file_path):
-    """
-    Tente de déchiffrer un message crypté avec un chiffrement de César en essayant tous les décalages possibles (1 à 25).
-    """
-    try:
-        with open(file_path, 'r') as file:
-            encrypted_message = file.read().strip()
-
-        print(f"Message crypté : {encrypted_message}")
-        print("\nTentatives de déchiffrement :\n")
-
-        for shift in range(1, 26):
-            decrypted_message = caesar_decrypt(encrypted_message, shift)
-            print(f"Décalage {shift} : {decrypted_message}")
-
-    except FileNotFoundError:
-        print(f"Erreur : le fichier '{file_path}' est introuvable.")
-    except Exception as e:
-        print(f"Une erreur est survenue : {e}")
+    if 'message1' in file_name:
+        return 'cesar'
+    elif 'message2' in file_name:
+        return 'vigenere'
+    elif 'message3' in file_name:
+        return 'adfgvx'
+    else:
+        return None
 
 if __name__ == "__main__":
     if len(sys.argv) != 2:
-        print("Utilisation : python3 main.py <nom_du_fichier>")
+        print("Utilisation : python3 src/main.py <nom_du_fichier>")
         sys.exit(1)
 
     file_path = sys.argv[1]
+    file_name = os.path.basename(file_path)
 
-    brute_force_caesar(file_path)
+    cipher_type = detect_cipher_type(file_name)
 
-def vigenere_decrypt(ciphertext, key):
-    """
-    Déchiffre un message chiffré avec le chiffrement de Vigenère en utilisant une clé donnée.
-    """
-    decrypted_text = []
-    key_length = len(key)
-    key_as_int = [ord(i) - ord('A') for i in key]  
-    ciphertext_int = [ord(i) - ord('A') for i in ciphertext if i.isalpha()]  
-
-    key_index = 0  
-
-    for char in ciphertext:
-        if char.isalpha():  
-            value = (ord(char) - ord('A') - key_as_int[key_index]) % 26
-            decrypted_text.append(chr(value + ord('A')))  
-            key_index = (key_index + 1) % key_length  
-        else:
-            decrypted_text.append(char)  
-
-    return ''.join(decrypted_text)
-
-def read_ciphertext(file_path):
-    """
-    Lit le contenu d'un fichier et retourne le texte chiffré.
-    """
-    try:
-        with open(file_path, 'r') as file:
-            return file.read().strip() 
-    except FileNotFoundError:
-        print(f"Erreur : le fichier '{file_path}' est introuvable.")
+    if cipher_type is None:
+        print("Erreur : Impossible de déterminer le type de chiffrement à partir du nom de fichier.")
+        print("Assurez-vous que le nom du fichier contient 'message1', 'message2', ou 'message3'.")
         sys.exit(1)
 
-if __name__ == "__main__":
-    if len(sys.argv) != 2:
-        print("Utilisation : python3 main.py <nom_du_fichier>")
-        sys.exit(1)
+    if cipher_type == "cesar":
+        brute_force_caesar(file_path)
 
-    file_path = sys.argv[1]
+    elif cipher_type == "vigenere":
+        key = "CINQ"
+        ciphertext = read_ciphertext(file_path)
+        decrypted_message = vigenere_decrypt(ciphertext, key)
+        print(f"Message déchiffré :\n{decrypted_message}")
 
-    key = "CINQ"
+    elif cipher_type == "adfgvx":
+        MESSAGE = read_ciphertext(file_path)
+        key = 'CRYPTO'
+        message = fon(MESSAGE, key)
+        res = ""
+        for j in range(48):
+            res += message[0][j] + message[3][j] + message[5][j] + message[2][j] + message[4][j] + message[1][j]
+        print(adfgvx(res))
 
-    ciphertext = read_ciphertext(file_path)
-
-    decrypted_message = vigenere_decrypt(ciphertext, key)
-    print(f"Message déchiffré : {decrypted_message}")
-
-
-
-
+    else:
+        print("Type de chiffrement non supporté.")
